# Add benchmark and optimize `ContainsDuplicates`