### Changelist
`ContainsDuplicates` accounts for over 10% of the time when placing vault orders on testnet. Optimize the runtime by around 50%, especially in the almost-always case of no duplicates. Therefore this should speed up placement of vault orders by around 5%.

### Test Plan
- Added benchmarks

```
Before
BenchmarkContainsDuplicates_True-4    	12939104	       459.6 ns/op	     114 B/op	       1 allocs/op
BenchmarkContainsDuplicates_False-4   	 5010406	      1185 ns/op	     385 B/op	       3 allocs/op

After
BenchmarkContainsDuplicates_True-4    	21472302	       271.8 ns/op	     192 B/op	       1 allocs/op
BenchmarkContainsDuplicates_False-4   	11044560	       534.5 ns/op	     208 B/op	       1 allocs/op
```